Add IDL example
Add an example for how the Interface Definition Language works. An
example of the real nsIPromptFactory interface is shown, alongside a the
generated header file. Some example JavaScript and C++ implementations
of the interface are given. An example of invocations from the interface
in JavaScript and C++ are also given.

All of the code is from the actual source (slighlty amended for clarity)
except the C++ implementation (only a JavaScript implementation is used
in reality).