Update examples screenshots
Updates the screenshots of the three example applications following
changes to the code. The main changes are that bespoke icons are now
used for the toolbar icons and the dom walk outlining now uses shadow
rather than a border, so that the elements on the page don't change size
as a result of the red outline being added.